separated code into dimension.py and measure.py
